+++ intelmq/lib/pipeline.py
@@ -2,10 +2,11 @@
 import time
 from itertools import chain
 
+import redis
+
 import intelmq.lib.exceptions as exceptions
 import intelmq.lib.pipeline
 import intelmq.lib.utils as utils
-import redis
 
 __all__ = ['Pipeline', 'PipelineFactory', 'Redis', 'Pythonlist']
 
@@ -41,7 +42,8 @@ class Pipeline(object):
     def set_queues(self, queues, queues_type):
         """
         :param queues: For source queue, it's just string.
-                    For destination queue, it can be one of the following: None or list or dict (of strings or lists, one of the key should be '_default')
+                    For destination queue, it can be one of the following:
+                    None or list or dict (of strings or lists, one of the key should be '_default')
 
         :param queues_type: "source" or "destination"